+++ Model/Config/Source/StyleStore.php
@@ -1,89 +1,31 @@
 <?php
+
 namespace Feedaty\Badge\Model\Config\Source;
 
-use Magento\Framework\Option\ArrayInterface;
-use Magento\Framework\App\Config\ScopeConfigInterface;
-use Feedaty\Badge\Model\Config\Source\WebService;
-use \Magento\Store\Model\StoreManagerInterface;
-use \Magento\Framework\App\Request\Http;
-use Feedaty\Badge\Helper\Data as DataHelp;
+use \Magento\Framework\Option\ArrayInterface;
+use Feedaty\Badge\Helper\WidgetHelper;
 
 class StyleStore implements ArrayInterface
 {
-    /**
-    * @var \Magento\Framework\App\Config\ScopeConfigInterface
-    */
-    protected $scopeConfig;
 
     /**
-    * @var Feedaty\Badge\Helper\Data
-    */
-    protected $dataHelper;
+     * @var WidgetHelper
+     */
+    protected $_widgetHelper;
 
-    /*
-    * Constructor
-    *
-    */
-    public function __construct(
-        ScopeConfigInterface $scopeConfig,
-        StoreManagerInterface $storeManager,
-        DataHelp $dataHelper,
-        Http $request,
-        WebService $fdservice
-        )
+    /**
+     * @param WidgetHelper $widgetHelper
+     */
+    public function __construct(WidgetHelper $widgetHelper)
     {
-        $this->scopeConfig = $scopeConfig;
-        $this->storeManager = $storeManager;
-        $this->_dataHelper = $dataHelper;
-        $this->_request = $request;
-        $this->_fdservice = $fdservice;
+        $this->_widgetHelper = $widgetHelper;
     }
 
     /**
-    *
-    * @return $return
-    */
-    public function toOptionArray() {
-
-        $return = array();
-
-        $store_scope = \Magento\Store\Model\ScopeInterface::SCOPE_STORE;
-
-        $store = $this->storeManager->getStore($this->_request->getParam('store', 0));
-
-        $merchant_code = $store->getConfig('feedaty_global/feedaty_preferences/feedaty_code');
-
-        if($this->_request->getParam('store', 0) == 0) {
-
-            $merchant_code = $this->scopeConfig->getValue('feedaty_global/feedaty_preferences/feedaty_code', $store_scope);
-
-        }
-
-        if (strlen($merchant_code == 0))  {
-
-            return $return;
-
-        }
-
-        $dataObject = $this->_fdservice->getFeedatyData($merchant_code);
-
-        $data = $dataObject['merchant']['variants'];
-
-        if($data) {
-
-            // get Badges
-            foreach ($data as $k => $v) {
-
-               // if ($v['type'] == "merchant" && $v['name'] != "dynamic" ) {
-              //  if ($v['type'] == "merchant"  ) {
-                    $return[] = ['value' => $k,'label' => $v['name_shown_it-IT']];
-
-             //   }
-
-            }
-
-        }
-
-        return $return;
+     * @return array
+     */
+    public function toOptionArray()
+    {
+        return $this->_widgetHelper->badgeData('merchant');
     }
 }
